+++ GR_Core.lua
@@ -190,10 +190,10 @@ end
 function core:LoadTables()
     ns.tblBlackList, ns.antiSpamList = {}, {}
 
-    local blSuccess, tblBL = ns.code:decompressData(ns.g.blackList or {})
+    local blSuccess, tblBL = ns.code:decompressData(ns.guild.blackList)
     ns.tblBlackList = blSuccess and tblBL or {}
 
-    local asSuccess, tblAS = ns.code:decompressData(ns.g.antiSpamList or {})
+    local asSuccess, tblAS = ns.code:decompressData(ns.guild.antiSpamList)
     ns.tblAntiSpamList = asSuccess and tblAS or {}
 
     --* Load Class/Race and Invalid Zones Table
@@ -240,7 +240,6 @@ function core:PerformRecordMaintenance()
     end
 
     removeOldRecords(ns.tblAntiSpamList, ns.gmSettings.antiSpamDays)
-    --removeOldRecords(ns.tblBlackList, ns.gmSettings.antiSpamDays)
 end
 function core:StartSlashCommands() -- Start Slash Commands
     local function slashCommand(msg)